// src/types/config.ts
import { Dispatch, SetStateAction } from 'react';
import { MenuOrientation, ThemeDirection } from 'config';

export type FontFamily = string;
export type PresetColor = 'default' | 'theme1' | 'theme2' | 'theme3' | 'theme4' | 'theme5' | 'theme6' | 'theme7' | 'theme8';
export type I18n = 'en' | 'fr' | 'ro' | 'zh' | 'ar'; // 'en' - English, 'fr' - French, 'ro' - Romanian, 'zh' - Chinese, 'ar' - Arabic

// ==============================|| CONFIG TYPES ||============================== //

export type ConfigStates = {
  /**
   * The props used for the theme font-style.
   * We use options from next/font ref -
   * https://nextjs.org/docs/pages/building-your-application/optimizing/fonts
   * https://fonts.google.com/variablefonts
   */
  fontFamily: FontFamily;

  /**
   * The props used for display menu-items with multi-language.
   * We provide static below languages according to 'react-intl' options - https://www.npmjs.com/package/react-intl
   * 'en' (default)
   * 'fr'
   * 'ro'
   * 'zh'
   */
  i18n: I18n;

  /**
   * the props used for menu orientation (diffrent theme layout).
   * we provide static below options -
   * 'vertical' (default) - MenuOrientation.VERTICAL
   * 'horizontal' - MenuOrientation.HORIZONTAL
   * 'mini-vertical' - MenuOrientation.MINI_VERTICAL
   */
  menuOrientation: MenuOrientation;

  /**
   * the props used for theme container.
   * the container centers your content horizontally. It's the most basic layout element.
   * default - true which show container
   * false - will show fluid
   */
  container: boolean;

  /**
   * the props used for theme primary color variants
   * we provide static below options thoe s are already defaine in src/themes/theme -
   * 'default'
   * 'theme1'
   * 'theme2'
   * 'theme3'
   * 'theme4'
   * 'theme5'
   * 'theme6'
   * 'theme7'
   * 'theme8'
   */
  presetColor: PresetColor;

  /**
   * the props used for default theme direction
   * explore the default theme
   * below theme options -
   * 'ltr' (default) - ThemeDirection.LTR
   * 'rtl' - ThemeDirection.RTL
   */
  themeDirection: ThemeDirection;
};

export type ConfigContextValue = {
  state: ConfigStates;
  setState: Dispatch<SetStateAction<ConfigStates>>;
  setField: (name: keyof ConfigStates, updateValue: ConfigStates[keyof ConfigStates]) => void;
  resetState: () => void;
};